% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyModel.R
\name{applyModel}
\alias{applyModel}
\title{Produce a segmentation based on a given model and extract enhancer / promoter elemenents.}
\usage{
applyModel(regions, model = NULL, provideModel = FALSE, genomeSize,
  counts = NULL, bamtab = NULL, outdir = ".", nthreads = 1,
  learnTrans = FALSE, refCounts = NULL)
}
\arguments{
\item{regions}{GRanges object containing the genomic regions of interest.}

\item{model}{A list with the parameters that describe the HMM.}

\item{provideModel}{flag, whether or not to use the provided model that was learned on mouse embryonic stem cell data.}

\item{genomeSize}{vector with chromosome lengths.}

\item{counts}{Count matrix matching with the \code{regions} parameter.
Each row of the matrix represents a mark and each column a bin resulting
from dividing the genomic regions into non-overlapping bins of equal size.
The rows of the matrix must be named with the name of the marks and these names must be unique.
If not given, it will be calculated and written to the output directory.}

\item{bamtab}{Data frame describing how to get the counts from each file. 
The following columns are required:
'mark' (name of the histone mark),
'path' (path to the bam file).}

\item{outdir}{path to the output directory.}

\item{nthreads}{number of threads used for learning.}

\item{learnTrans}{flag, whether or not to let the model learn the transition probabilities while fixing the emission parameters.}

\item{refCounts}{Count matrix of the reference model.
Has to be given if \code{refCounts} and \code{counts} have different dimensions.}
}
\value{
nothing.
}
\description{
Produce a segmentation based on a given model and extract enhancer / promoter elemenents.
}
