% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialScores.R
\name{differentialScores}
\alias{differentialScores}
\title{Produce a segmentation based on a given model and extract enhancer / promoter elemenents.}
\usage{
differentialScores(scores1, scores2, outfile = "differentialScores.bed")
}
\arguments{
\item{scores1}{GRanges object containing scores.}

\item{scores2}{GRanges object containing scores.}

\item{outfile}{path to the output file.}
}
\value{
nothing.
}
\description{
Produce a segmentation based on a given model and extract enhancer / promoter elemenents.
}
