% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{ehmm}
\alias{ehmm}
\alias{ehmm-package}
\title{Chromatin segmentation}
\description{
The package provides methods for finding chromatin states in 
epigenomic data. There is an R interface as well as a 
command line interface.
}
\author{
Tobias Zehnder \email{zehnder@molgen.mpg.de}
}
