% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfoots.R
\name{fitNB}
\alias{fitNB}
\title{Fit a negative binomial distribution}
\usage{
fitNB(counts, posteriors = NULL, old_r = NULL, tol = 1e-08,
  verbose = FALSE, nthreads = 1)
}
\arguments{
\item{counts}{a vector of counts. If a list is given, then it is assumed 
to be the result of the function \code{mapToUnique(counts)}}

\item{posteriors}{a vector specifying a weight for each count. The maximized
function is: \eqn{\sum_{i=1}{L}{posteriors[i]\log(Prob\{counts[i]\}}}. If
not specified, equal weights will be assumed}

\item{old_r}{an initial value for the size parameter of the negative binomial.
If not specified the methods of moments will be used for an initial guess.}

\item{tol}{numerical tolerance of the fitting algorithm}

\item{nthreads}{number of threads. Too many threads might worsen the 
performance}
}
\value{
A list with the parameters of the negative binomial.
    \item{mu}{the mu parameter}
       \item{r}{the size parameter}
}
\description{
Maximum Likelihood Estimate for the parameters of a negative binomial distribution
generating a specified vector of counts. The MLE for the negative binomial
should not be used with a small number of datapoints, it is known to be
biased. Internally, this function is using Brent's method to find the
optimal dispersion parameter.
}
