% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnModel.R
\name{learnModel}
\alias{learnModel}
\title{Learn a model and produce a segmentation}
\usage{
learnModel(regions, nstates, bamtab, outdir = ".", nthreads = 1,
  pseudoCount = 1)
}
\arguments{
\item{regions}{GRanges object containing the genomic regions of interest.}

\item{nstates}{Number of states to learn.}

\item{bamtab}{Data frame describing how to get the counts from each file. 
The following columns are required:
'mark' (name of the histone mark),
'path' (path to the bam file).}

\item{outdir}{path to the output directory.}

\item{nthreads}{number of threads used for learning.}

\item{pseudoCount}{pseudo-count to add to read counts.
This is necessary because log-counts are calculated in order to fit a log-normal distribution.}
}
\value{
A list with the following arguments:
}
\description{
Learn a model and produce a segmentation
}
