% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{viterbi}
\alias{viterbi}
\title{Viterbi algorithm}
\usage{
viterbi(vscores, initP, trans, lliks, seqlens, endstate)
}
\arguments{
\item{initP}{matrix of initial probabilities: each column corresponds to a sequence}

\item{trans}{transition matrix (rows are previous state, columns are next state)}

\item{lliks}{matrix with emission probabilities for each datapoint and each state.
Columns are datapoints and rows are states.}

\item{seqlens}{length of each subsequence of datapoints (set this to ncol(lliks)
if there is only one sequence).}
}
\value{
a list with the following arguments:
   \item{vpath}{viterbi path}
   \item{vllik}{log-likelihood of the viterbi path}
   \item{vscores}{viterbi scores}
}
\description{
Standard viterbi algorithm in the log space
}
